Completed Task 3 and code refactoring
## Task 3
Imagine a driver management frontend that is used internally by mytaxi
employees to create and edit driver related data.
For a new search functionality, we need an endpoint to search for
drivers.
It should be possible to search for drivers by their attributes
(username, online_status) as well as car characteristics (license plate,
rating, etc).

* implement a new endpoint for searching or extend an existing one
* driver/car attributes as input parameters
* return list of drivers